processor/requirements.txt
asyncpg==0.29.0
nats-py==2.6.0
boto3==1.29.7
httpx==0.25.2
sentence-transformers==2.2.2
torch==2.1.0